Move files for special numbers to own directory
