+++ maths/special_numbers/automorphic_number.py
