+++ maths/special_numbers/bell_numbers.py
