+++ maths/special_numbers/carmichael_number.py
