+++ maths/special_numbers/hamming_numbers.py
