+++ maths/special_numbers/hexagonal_number.py
