+++ maths/special_numbers/krishnamurthy_number.py
