+++ maths/special_numbers/perfect_number.py
