+++ maths/special_numbers/polygonal_numbers.py
