+++ maths/special_numbers/pronic_number.py
