+++ maths/special_numbers/proth_number.py
