+++ maths/special_numbers/weird_number.py
